Simplify account page
