+++ src/pages/AccountPage.js
@@ -1,54 +1,23 @@
 import { useState, useEffect, useContext } from 'react';
-import { Box, Typography, Button, Container, Divider, Grid, Card, TextField, List, ListItem, ListItemIcon, ListItemText, IconButton, Avatar, Chip, Skeleton, LinearProgress, CircularProgress } from '@mui/material';
-import { Receipt, Download, Edit, Save, Cancel, Block, SupportAgent, Bolt, AutoFixHighRounded } from '@mui/icons-material';
-import { API, Auth, Storage } from 'aws-amplify';
-import { LoadingButton } from '@mui/lab';
+import { Box, Typography, Button, Container, Divider, Grid, Card, List, ListItem, ListItemIcon, ListItemText, IconButton, Chip, Skeleton, LinearProgress } from '@mui/material';
+import { Receipt, Download, Block, SupportAgent, Bolt, AutoFixHighRounded } from '@mui/icons-material';
+import { API } from 'aws-amplify';
 import { UserContext } from '../UserContext';
 import { useSubscribeDialog } from '../contexts/useSubscribeDialog';
-import { SnackbarContext } from '../SnackbarContext';
 
 const AccountPage = () => {
   const userDetails = useContext(UserContext);
   const { openSubscriptionDialog } = useSubscribeDialog();
-  const [email, setEmail] = useState(userDetails?.user?.attributes?.email || '');
-  const [editingEmail, setEditingEmail] = useState(false);
-  const [verificationCode, setVerificationCode] = useState('');
-  const [isVerifyingEmail, setIsVerifyingEmail] = useState(false);
-  const [currentPassword, setCurrentPassword] = useState('');
-  const [newPassword, setNewPassword] = useState('');
-  const [confirmPassword, setConfirmPassword] = useState('');
-  const [isPasswordValid, setIsPasswordValid] = useState(false);
-  const [passwordError, setPasswordError] = useState('');
-  const [updatingPassword, setUpdatingPassword] = useState(false);
-  const [profilePhoto, setProfilePhoto] = useState(userDetails?.user?.profilePhoto || '');
-  const [uploadingProfilePhoto, setUploadingProfilePhoto] = useState(false);
   const [invoices, setInvoices] = useState([]);
   const [hasMore, setHasMore] = useState(false);
   const [loadingInvoices, setLoadingInvoices] = useState(true);
   const [page, setPage] = useState(1);
   const [loadingSubscription, setLoadingSubscription] = useState(true);
-  const { setOpen, setSeverity, setMessage } = useContext(SnackbarContext);
-  const [updatingEmail, setUpdatingEmail] = useState(false);
 
   useEffect(() => {
     fetchInvoices();
     fetchSubscription();
   }, [page]);
-
-  useEffect(() => {
-    setEmail(userDetails?.user?.attributes?.email);
-    console.log(userDetails)
-  }, [userDetails]);
-
-  useEffect(() => {
-    if (newPassword === confirmPassword) {
-      setIsPasswordValid(newPassword.length >= 8);
-      setPasswordError('');
-    } else {
-      setIsPasswordValid(false);
-      setPasswordError('Passwords do not match');
-    }
-  }, [newPassword, confirmPassword]);
 
   const fetchInvoices = async () => {
     try {
@@ -88,109 +57,6 @@ const AccountPage = () => {
     window.open(url, '_blank');
   };
 
-  const handleUpdateEmail = async () => {
-    try {
-      setUpdatingEmail(true);
-      const user = await Auth.currentAuthenticatedUser();
-      await Auth.updateUserAttributes(user, { email });
-      setEditingEmail(false);
-      setIsVerifyingEmail(true);
-      setMessage('Verification code sent to your email. Please enter the code to verify your new email address.');
-      setSeverity('info');
-      setOpen(true);
-    } catch (error) {
-      console.error('Error updating email:', error);
-      setMessage('Something went wrong. Please try again.');
-      setSeverity('error');
-      setOpen(true);
-    } finally {
-      setUpdatingEmail(false);
-    }
-  };
-
-  const handleVerifyEmail = async () => {
-    try {
-      await Auth.verifyCurrentUserAttributeSubmit('email', verificationCode);
-      setIsVerifyingEmail(false);
-      setVerificationCode('');
-      setMessage('Email address verified successfully!');
-      setSeverity('success');
-      setOpen(true);
-    } catch (error) {
-      console.error('Error verifying email:', error);
-      setMessage('Invalid verification code. Please try again.');
-      setSeverity('error');
-      setOpen(true);
-    }
-  };
-
-  const handleUpdatePassword = async (event) => {
-    event.preventDefault();
-
-    if (!isPasswordValid) {
-      setPasswordError('Passwords do not match');
-      return;
-    }
-
-    try {
-      setUpdatingPassword(true);
-      const user = await Auth.currentAuthenticatedUser();
-      await Auth.changePassword(user, currentPassword, newPassword);
-      setCurrentPassword('');
-      setNewPassword('');
-      setConfirmPassword('');
-      setPasswordError('');
-    } catch (error) {
-      console.error('Error updating password:', error);
-      setPasswordError('Failed to update password. Please try again.');
-      setMessage('Something went wrong. Please try again.')
-      setSeverity('error')
-      setOpen(true)
-    } finally {
-      setUpdatingPassword(false);
-      setMessage('Your password has been updated!')
-      setSeverity('success')
-      setOpen(true)
-    }
-  };
-
-  const handleProfilePhotoUpload = async (event) => {
-    setUploadingProfilePhoto(true)
-    console.log(userDetails)
-    if (userDetails?.user?.sub) {
-      const file = event.target.files[0];
-      try {
-        // Upload the profile photo to Storage
-        const response = await Storage.put(`profilePictures/${userDetails?.user?.sub}`, file, {
-          contentType: file.type, // contentType is optional
-        });
-
-        // Get the current authenticated user
-        const userObj = await Auth.currentAuthenticatedUser();
-
-        // Update the user's profile with the uploaded photo key
-        const updatePictureResponse = await Auth.updateUserAttributes(userObj, {
-          picture: response.key
-        });
-
-        const newProfilePicture = await Storage.get(response.key)
-        userDetails?.setUser({
-          ...userDetails.user,
-          profilePhoto: newProfilePicture
-        })
-        console.log(newProfilePicture)
-
-        console.log(updatePictureResponse);
-        setUploadingProfilePhoto(false)
-      } catch (error) {
-        console.error('Error uploading profile photo:', error);
-        setUploadingProfilePhoto(false)
-      }
-      console.log('Uploading profile photo:', file);
-      setUploadingProfilePhoto(false)
-    }
-  };
-
   const isLoading = loadingInvoices || loadingSubscription;
   const recentPaidInvoice = invoices.find((invoice) => invoice.paid);
   const currentSubscription = recentPaidInvoice?.lines?.data?.[0]?.description
@@ -204,7 +70,7 @@ const AccountPage = () => {
       </Typography>
       <Divider sx={{ my: 3 }} />
 
-      <Grid container spacing={3}>
+      <Grid container spacing={3} alignItems="flex-start">
         <Grid item xs={12} md={6}>
           <Card sx={{ p: 3, backgroundColor: 'background.paper' }}>
             <Typography variant="h5" gutterBottom>
@@ -297,128 +163,8 @@ const AccountPage = () => {
             </Box>
           </Card>
         </Grid>
-        <Grid item xs={12} md={6}>
-          <Card sx={{ p: 3, backgroundColor: 'background.paper' }}>
-            <Typography variant="h5" gutterBottom>
-              Account Information
-            </Typography>
-            <Box sx={{ display: 'flex', alignItems: 'center', mb: uploadingProfilePhoto ? 1 : 4 }}>
-              <Avatar src={userDetails?.user?.profilePhoto || null} sx={{ width: 80, height: 80, mr: 3 }} />
-              <Button variant="contained" component="label" disabled={uploadingProfilePhoto}>
-                Upload Photo
-                <input type="file" hidden onChange={handleProfilePhotoUpload} accept="image/*" />
-              </Button>
-            </Box>
-            {uploadingProfilePhoto &&
-              <Box>
-                <Typography textAlign='center' fontSize={14} mb={1}>
-                  Uploading...
-                </Typography>
-                <LinearProgress sx={{ mb: 4 }} />
-              </Box>
-            }
-            <Box sx={{ mb: 3 }}>
-            <TextField
-              label="Email"
-              autoComplete="off"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              fullWidth
-              disabled={!editingEmail || isVerifyingEmail || updatingEmail}
-              InputProps={{
-                endAdornment: editingEmail ? (
-                  <>
-                    <IconButton onClick={handleUpdateEmail} disabled={isVerifyingEmail || updatingEmail}>
-                      {updatingEmail ? <CircularProgress size={24} /> : <Save />}
-                    </IconButton>
-                    <IconButton
-                      onClick={() => {
-                        setEmail(userDetails?.user?.userDetails?.email || '');
-                        setEditingEmail(false);
-                      }}
-                      disabled={isVerifyingEmail || updatingEmail}
-                    >
-                      <Cancel />
-                    </IconButton>
-                  </>
-                ) : (
-                  <IconButton onClick={() => setEditingEmail(true)} disabled={isVerifyingEmail || updatingEmail}>
-                    <Edit />
-                  </IconButton>
-                ),
-              }}
-            />
-              {isVerifyingEmail && (
-                <Box sx={{ mt: 2 }}>
-                  <TextField
-                    label="Verification Code"
-                    value={verificationCode}
-                    onChange={(e) => setVerificationCode(e.target.value)}
-                    fullWidth
-                    sx={{ mb: 2 }}
-                  />
-                  <Button variant="contained" onClick={handleVerifyEmail}>
-                    Verify Email
-                  </Button>
-                </Box>
-              )}
-            </Box>
-            <Box sx={{ mb: 3 }}>
-              <form onSubmit={handleUpdatePassword}>
-                <TextField
-                  label="Username"
-                  type="text"
-                  autoComplete="username"
-                  value={userDetails?.user?.username || ''}
-                  fullWidth
-                  disabled
-                  sx={{ mb: 2 }}
-                />
-                <TextField
-                  label="Current Password"
-                  type="password"
-                  autoComplete="current-password"
-                  value={currentPassword}
-                  onChange={(e) => setCurrentPassword(e.target.value)}
-                  fullWidth
-                  sx={{ mb: 2 }}
-                />
-                <TextField
-                  label="New Password"
-                  type="password"
-                  autoComplete="new-password"
-                  value={newPassword}
-                  onChange={(e) => setNewPassword(e.target.value)}
-                  fullWidth
-                  sx={{ mb: 2 }}
-                  error={!!passwordError}
-                  helperText={passwordError}
-                />
-                <TextField
-                  label="Confirm New Password"
-                  type="password"
-                  autoComplete="new-password"
-                  value={confirmPassword}
-                  onChange={(e) => setConfirmPassword(e.target.value)}
-                  fullWidth
-                  error={!!passwordError}
-                  helperText={passwordError}
-                />
-                <LoadingButton
-                  type="submit"
-                  variant="contained"
-                  disabled={!isPasswordValid || updatingPassword}
-                  sx={{ mt: 3 }}
-                  loading={updatingPassword}
-                >
-                  {updatingPassword ? 'Updating...' : 'Update Password'}
-                </LoadingButton>
-              </form>
-            </Box>
-          </Card>
-        </Grid>
 
-        <Grid item xs={12}>
+        <Grid item xs={12} md={6}>
           <Card sx={{ p: 3, backgroundColor: 'background.paper' }}>
             <Typography variant="h5" gutterBottom>
               Invoices
@@ -447,7 +193,7 @@ const AccountPage = () => {
               </Button>
             )}
             {loadingInvoices ? (
-              <>
+              <> 
                 {[...Array(3)].map((_, index) => (
                   <Skeleton key={index} variant="rectangular" height={60} sx={{ borderRadius: 1, mb: 1 }} />
                 ))}
